Update ProfilesController to validate user
Ember is sending user_id along with the req to update Profile
so instead of validating if current_user == current_user (which
was useless), I'm not validating that current_user.id == user_id
that was sent in the req. Still not sure if this is the best way
to handle it, but at least I'm doing more validation than I was.