Change positions to be bools instead of an array
Previously positions was being represented by an array of strings.
I decided to make each position a bool on the profile object instead
of making an array like ['carry', 'support', etc] because this allows
Emberizing this data to be a little easier. It also helps me prevent
requests with ANY text from being processed.
Other options: go back to array data type and force the values to be one
of the positions ('carry', 'offlaner', etc.), and make sure people
aren't passing strings that don't correlate to a position ('dog').
Issues with this then become location in the array? Is 'carry' always
positions[0]? What if a user doesn't play the 'carry' roll. I didn't
like relying on the right position being in the right place in the
array so I added them each as bools. I might wrap this up into its
own model in the future, or investigate the trade-offs I've decided on
but for now it works.