Adds comment polymorphic relationship to profile & game
This allows a user to create a comment that belongs to either a profile
or a game. Attempted this using a poylmorphic relationship so that a
user can create a comment for a user's profile as a whole, or a
specific game. I'm not sure this is the best solution, but wanted to
give it a shot to see if it would work. Was successful testing creation
and indexing on comments that belong to a user's profile ONLY which
seems to be working currently, but is certainly a work in progress.
Need to test full CRUD on these and make sure it's working on a game
as expected.